idempotent nix installation

+++ setup_ci.sh
@@ -5,7 +5,13 @@ set -e
 # set -x
 
 # Install nix
-curl --proto '=https' --tlsv1.2 -sSf -L https://install.determinate.systems/nix | sh -s -- install linux --no-confirm --init none
+
+if [ -f /nix/receipt.json ]; then
+    echo "nix is already installed"
+else
+    curl --proto '=https' --tlsv1.2 -sSf -L https://install.determinate.systems/nix | sh -s -- install linux --no-confirm --init none
+fi
+
 
 # Install devbox
 curl -fsSL https://get.jetpack.io/devbox | FORCE=1 bash
